proposed a refactor of selectCurrentType
