+++ src/selectors/currentTypeLabel.js
@@ -1,0 +1,9 @@
+import get from 'lodash.get'
+import { createSelector } from 'reselect'
+
+import selectFormOccasion from './formOccasion'
+
+export default createSelector(
+  selectFormOccasion,
+  formOccasion => get(formOccasion, 'type')
+)
